improve code splitting
