+++ frontend/src/App.js
@@ -1,6 +1,7 @@
-import React from 'react';
+import React, { Suspense } from 'react';
 import { Router, Redirect, globalHistory } from '@reach/router';
 import { QueryParamProvider } from 'use-query-params';
+import ReactPlaceholder from 'react-placeholder';
 import { useMeta } from 'react-meta-elements';
 import { connect } from 'react-redux';
 import * as Sentry from '@sentry/react';
@@ -41,7 +42,6 @@ import { NotFound } from './views/notFound';
 import { SelectTask } from './views/taskSelection';
 import { MapTask, ValidateTask } from './views/taskAction';
 import { EmailVerification } from './views/verifyEmail';
-import { ProjectEdit } from './views/projectEdit';
 import { ProjectStats } from './views/projectStats';
 import { ContactPage } from './views/contact';
 import { SwaggerView } from './views/swagger';
@@ -53,8 +53,10 @@ import {
 } from './views/notifications';
 import { Banner } from './components/banner/index';
 
-/*TODO(tdk): if QueryParamProvider is not needed elsewhere,
- *  create special sub-router for Projects page and wrap it only around that */
+const ProjectEdit = React.lazy(() =>
+  import('./views/projectEdit' /* webpackChunkName: "projectEdit" */),
+);
+
 let App = (props) => {
   useMeta({ property: 'og:url', content: process.env.REACT_APP_BASE_URL });
   useMeta({ name: 'author', content: ORG_NAME });
@@ -70,68 +72,72 @@ let App = (props) => {
             <Header path="/*" />
           </Router>
           <main className="cf w-100 base-font">
-            <QueryParamProvider reachHistory={globalHistory}>
-              <Router primary={false}>
-                <Home path="/" />
-                <ProjectsPage path="explore">
-                  <ProjectsPageIndex path="/" />
-                  <MoreFilters path="/filters/*" />
-                </ProjectsPage>
-                <ProjectDetailPage path="projects/:id" />
-                <SelectTask path="projects/:id/tasks" />
-                <MapTask path="projects/:id/map" />
-                <ValidateTask path="projects/:id/validate" />
-                <ProjectStats path="projects/:id/stats" />
-                <OrganisationStats path="organisations/:id/stats/" />
-                <LearnPage path="learn" />
-                <QuickstartPage path="learn/quickstart" />
-                <AboutPage path="about" />
-                <ContactPage path="contact/" />
-                <ContributionsPageIndex path="contributions">
-                  <UserStats path="/" />
-                  <ContributionsPage path="tasks/*" />
-                  <UserProjectsPage path="projects/*" />
-                  <MyTeams path="teams/*" />
-                </ContributionsPageIndex>
-                <UserDetail path="users/:username" />
-                <NotificationsPage path="inbox">
-                  <NotificationPageIndex path="/" />
-                  <NotificationDetail path="message/:id" />
-                </NotificationsPage>
-                <Authorized path="authorized" />
-                <Login path="login" />
-                <Welcome path="welcome" />
-                <Settings path="settings" />
-                <EmailVerification path="verify-email" />
-                <ManagementSection path="manage">
-                  <ManagementPageIndex path="/" />
-                  <ListOrganisations path="organisations/" />
-                  <CreateOrganisation path="organisations/new/" />
-                  <EditOrganisation path="organisations/:id/" />
-                  <ManageTeams path="teams/" />
-                  <UsersList path="users/" />
-                  <CreateTeam path="teams/new" />
-                  <EditTeam path="teams/:id" />
-                  <ListCampaigns path="campaigns/" />
-                  <CreateCampaign path="campaigns/new" />
-                  <EditCampaign path="campaigns/:id" />
-                  <CreateProject path="projects/new" />
-                  <ProjectEdit path="projects/:id" />
-                  <ManageProjectsPage path="projects/*" />
-                  <ListInterests path="categories/" />
-                  <EditInterest path="categories/:id" />
-                  <CreateInterest path="categories/new" />
-                  <CreateLicense path="licenses/new" />
-                  <ListLicenses path="licenses/" />
-                  <EditLicense path="licenses/:id" />
-                </ManagementSection>
-                <TeamDetail path="teams/:id/membership" />
-                <SwaggerView path="/api-docs/" />
-                <Redirect from="project/:id" to="/projects/:id" noThrow />
-                <Redirect from="contribute/" to="/explore" noThrow />
-                <NotFound default />
-              </Router>
-            </QueryParamProvider>
+            <Suspense
+              fallback={<ReactPlaceholder showLoadingAnimation={true} rows={30} delay={300} />}
+            >
+              <QueryParamProvider reachHistory={globalHistory}>
+                <Router primary={false}>
+                  <Home path="/" />
+                  <ProjectsPage path="explore">
+                    <ProjectsPageIndex path="/" />
+                    <MoreFilters path="/filters/*" />
+                  </ProjectsPage>
+                  <ProjectDetailPage path="projects/:id" />
+                  <SelectTask path="projects/:id/tasks" />
+                  <MapTask path="projects/:id/map" />
+                  <ValidateTask path="projects/:id/validate" />
+                  <ProjectStats path="projects/:id/stats" />
+                  <OrganisationStats path="organisations/:id/stats/" />
+                  <LearnPage path="learn" />
+                  <QuickstartPage path="learn/quickstart" />
+                  <AboutPage path="about" />
+                  <ContactPage path="contact/" />
+                  <ContributionsPageIndex path="contributions">
+                    <UserStats path="/" />
+                    <ContributionsPage path="tasks/*" />
+                    <UserProjectsPage path="projects/*" />
+                    <MyTeams path="teams/*" />
+                  </ContributionsPageIndex>
+                  <UserDetail path="users/:username" />
+                  <NotificationsPage path="inbox">
+                    <NotificationPageIndex path="/" />
+                    <NotificationDetail path="message/:id" />
+                  </NotificationsPage>
+                  <Authorized path="authorized" />
+                  <Login path="login" />
+                  <Welcome path="welcome" />
+                  <Settings path="settings" />
+                  <EmailVerification path="verify-email" />
+                  <ManagementSection path="manage">
+                    <ManagementPageIndex path="/" />
+                    <ListOrganisations path="organisations/" />
+                    <CreateOrganisation path="organisations/new/" />
+                    <EditOrganisation path="organisations/:id/" />
+                    <ManageTeams path="teams/" />
+                    <UsersList path="users/" />
+                    <CreateTeam path="teams/new" />
+                    <EditTeam path="teams/:id" />
+                    <ListCampaigns path="campaigns/" />
+                    <CreateCampaign path="campaigns/new" />
+                    <EditCampaign path="campaigns/:id" />
+                    <CreateProject path="projects/new" />
+                    <ProjectEdit path="projects/:id" />
+                    <ManageProjectsPage path="projects/*" />
+                    <ListInterests path="categories/" />
+                    <EditInterest path="categories/:id" />
+                    <CreateInterest path="categories/new" />
+                    <CreateLicense path="licenses/new" />
+                    <ListLicenses path="licenses/" />
+                    <EditLicense path="licenses/:id" />
+                  </ManagementSection>
+                  <TeamDetail path="teams/:id/membership" />
+                  <SwaggerView path="/api-docs/" />
+                  <Redirect from="project/:id" to="/projects/:id" noThrow />
+                  <Redirect from="contribute/" to="/explore" noThrow />
+                  <NotFound default />
+                </Router>
+              </QueryParamProvider>
+            </Suspense>
           </main>
           {MATOMO_ID && <Banner />}
           <Router primary={false}>
