+++ frontend/src/components/projectDetail/index.js
@@ -22,7 +22,7 @@ import { useSetProjectPageTitleTag } from '../../hooks/UseMetaTags';
 import { useFetch } from '../../hooks/UseFetch';
 
 /* lazy imports must be last import */
-const ProjectTimeline = React.lazy(() => import('./timeline'));
+const ProjectTimeline = React.lazy(() => import('./timeline' /* webpackChunkName: "timeline" */));
 
 const ProjectDetailMap = (props) => {
   const [taskBordersOnly, setTaskBordersOnly] = useState(true);
@@ -83,7 +83,7 @@ export const ProjectDetailLeft = ({ project, contributors, className, type }: Ob
       <div className="h-75 z-1">
         <ReactPlaceholder
           showLoadingAnimation={true}
-          rows={3}
+          rows={10}
           delay={500}
           ready={typeof project.projectId === 'number'}
         >
