+++ frontend/src/components/userDetail/tests/editByNumbers.test.js
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 
 import { ReduxIntlProviders } from '../../../utils/testWithIntl';
-import { EditsByNumbers } from '../editsByNumbers';
+import EditsByNumbers from '../editsByNumbers';
 
 jest.mock('react-chartjs-2', () => ({
   Doughnut: () => null,
